Bypass the default setting of CPP in OS X and explicitly use /usr/bin/cpp. This makes f2c work.  Bump PKGREVISION.
The default CPP points to gcc, which broke f2c because gcc is unable to
preprocess Fortran files as the Fortran compiler is not installed.
Surprisingly, /usr/bin/cpp is just a wrapper around gcc, but it explicitly
passes '-x c' to cpp to force the language to be C and bypass the .F
extension.