Update to 8:
The current version is release 8 of 10-Jan-2010. Version 8 of the
JPEG software introduces basic new features for image coding
application and comes with an essential update of the original JPEG
specification.

New features and enhancements have been developed for the three
major parts of the IJG library: the compression, the decompression,
and the transcoding parts. These parts are represented by the three
major tools of the IJG software: cjpeg (compress JPEG), djpeg
(decompress JPEG), and jpegtran (transcode JPEG - lossless):  IJG
Tools New jpegtran -scale lossless resize feature

A new option has been added to the jpegtran tool and transcoding
library:

	-scale M/N     Scale output image by fraction M/N, eg, 1/8

Currently supported scale factors are M/N with all M from 1 to 16,
where N is the source DCT size, which is 8 for baseline JPEG. If
the /N part is omitted, then M specifies the DCT scaled size to be
applied on the given input. For baseline JPEG this is equivalent
to M/8 scaling, since the source DCT size for baseline JPEG is 8.
CAUTION: An implementation of the JPEG SmartScale extension is
required for this feature. SmartScale enabled JPEG is not yet widely
implemented, so many decoders will be unable to view a SmartScale
extended JPEG file at all.

New djpeg -scale feature

The -scale option of the djpeg tool and decompression library has
been extended:

	-scale M/N     Scale output image by fraction M/N, eg, 1/8

Currently supported scale factors are M/N with all M from 1 to 16,
where N is the source DCT size, which is 8 for baseline JPEG. If
the /N part is omitted, then M specifies the DCT scaled size to be
applied on the given input. For baseline JPEG this is equivalent
to M/8 scaling, since the source DCT size for baseline JPEG is 8.

New cjpeg features

Two new features have been added to the cjpeg tool and compression
library:

	-quality N[,...]   Compression quality (0..100; 5-95 is
	useful range) -scale M/N     Scale image by fraction M/N,
	eg, 1/2

The -quality option has been extended for support of separate
quality settings for luminance and chrominance (or in general, for
every provided quantization table slot). This feature is useful
for high-quality applications which cannot accept the damage of
color data by coarse subsampling settings. You can now easily reduce
the color data amount more smoothly with finer control without
separate subsampling. The resulting file is fully compliant with
standard JPEG decoders.

A new -scale option is provided with cjpeg which complements the
corresponding djpeg -scale option. The supported range of scaling
factors is 8/N with all N=1...16.  This means you can now easily
alter the nominal spatial resolution of a given source image while
compressing to JPEG without additional resampling.  For example,
if you have an image sensor providing an effective capture resolution
of 2268x1512 pixels (HI resolution), you can now directly generate
a MED resolution of 1512x1008 pixels (-scale 2/3) and a LOW resolution
of 1134x756 pixels (-scale 1/2) from the sensor source resolution
with the library while compressing to JPEG without additional
resampling.  (An efficient 12x12 FDCT is used in the -scale 2/3
case, and an efficient 16x16 FDCT is used in the -scale 1/2 case
instead of the standard 8x8 FDCT inside the library, ensuring
high-quality downscaled results - the resulting file is fully
compliant with standard JPEG decoders.)