Update to 1.6.2:
Changes: Two new methods of searching through the HTML documentation were
added: via a PHP based Web server and via an Eclipse help plugin. The
\addindex command now also works for the HTML indices. Rendering of LaTeX
formulas in the HTML output has been improved (especially for non-white
backgrounds), along with a number of smaller improvements and a large
number of bugfixes.